refactoring db con
